backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-cors==5.0.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
